+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/AppDaoAccess.kt
@@ -1,8 +1,9 @@
-package com.yoesuv.infomalangbatu.databases.place
+package com.yoesuv.infomalangbatu.databases
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import com.yoesuv.infomalangbatu.databases.place.PlaceRoom
 
 @Dao
 interface AppDaoAccess {
