+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/AppDatabase.kt
@@ -1,7 +1,8 @@
-package com.yoesuv.infomalangbatu.databases.place
+package com.yoesuv.infomalangbatu.databases
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.yoesuv.infomalangbatu.databases.place.PlaceRoom
 
 @Database(entities = [(PlaceRoom::class)], version =1)
 abstract class AppDatabase : RoomDatabase() {
