+++ app/src/main/java/com/yoesuv/infomalangbatu/databases/place/DatabaseInsertPlace.kt
@@ -1,6 +1,7 @@
 package com.yoesuv.infomalangbatu.databases.place
 
 import android.os.AsyncTask
+import com.yoesuv.infomalangbatu.databases.AppDatabase
 
 class DatabaseInsertPlace(private val appDatabase: AppDatabase, private val placeRoom: PlaceRoom): AsyncTask<Void, Void, Void>() {
